refactor: extract conda version handling into its own crate
