+++ Cargo.toml
@@ -219,6 +219,7 @@ path_resolver = { path = "crates/path_resolver", version = "=0.2.13", default-fe
 rattler = { path = "crates/rattler", version = "=0.48.4", default-features = false }
 rattler_cache = { path = "crates/rattler_cache", version = "=0.10.7", default-features = false }
 rattler_conda_types = { path = "crates/rattler_conda_types", version = "=0.51.0", default-features = false }
+rattler_conda_version = { path = "crates/rattler_conda_version", version = "=0.1.0", default-features = false }
 rattler_config = { path = "crates/rattler_config", version = "=0.7.2", default-features = false }
 rattler_digest = { path = "crates/rattler_digest", version = "=1.3.2", default-features = false }
 rattler_git = { path = "crates/rattler_git", version = "=0.3.2", default-features = false }
